refactor: remove babel
Switch to using `require` and `module.exports` instead of `import` and
`export default`.
