+++ .gitignore
@@ -1,5 +1,3 @@
 coverage/
 node_modules/
-rules/
-!docs/rules/
 .nyc_output/
